Move FWB js from apps to routes js
