+++ cfgov/unprocessed/js/routes/consumer-tools/financial-well-being/index.js
@@ -1,2 +1,2 @@
-import fwbQuestions from '../../../../apps/financial-well-being/js/fwb-questions';
+import fwbQuestions from './fwb-questions.js';
 fwbQuestions.init();
