+++ cfgov/unprocessed/js/routes/consumer-tools/financial-well-being/results/index.js
@@ -1,2 +1,2 @@
-import fwbResults from '../../../../../apps/financial-well-being/js/fwb-results';
+import fwbResults from '../fwb-results.js';
 fwbResults.init();
